chore: update npm dependency to latest stable version (#2145)
* chore: update npm dependency to latest stable version

* chore: update npm dependency to latest stable version

* docs(changeset, readme): update changeset instructions and improve readme

The changeset README has been updated with detailed instructions on how to write a changeset, including examples and guidelines for choosing version tags.

The main project README has been improved with a better overview of the Fusion Framework, sections on getting started, cookbooks, and contributing. This provides more context and guidance for new users and contributors.