fix(http)!: throw correct error in jsonSelector (#2196)
The `jsonSelector` function is not checking the error type in the `catch`
block. This led to always throwing a parser error, even when the actual
error was due to an unsuccessful response.

BREAKING CHANGE:
If for some reason developers has catched the error and assumed the cause property would give the proper error data, this will no longer be the case.