docs(module-http): Update documentation for http module
This commit updates the documentation in `README.md` for the http module. It includes an introduction to the module, a concepts section highlighting key concepts, a sequence diagram for http request execution, examples, and improved documentation for configuring http module and working with http clients. The formatting of the documentation has also been improved.