Refactor/module http/improve doc (#2324)
* refactor(module-http): update response selector function signature

Update the signature of the `ResponseSelector` function in the `types.ts` file of the `http` module. This change allows for more flexibility in the types of the response and the result data.

* docs(module-http): update tsdoc

* docs(module-http): Update documentation for http module

This commit updates the documentation in `README.md` for the http module. It includes an introduction to the module, a concepts section highlighting key concepts, a sequence diagram for http request execution, examples, and improved documentation for configuring http module and working with http clients. The formatting of the documentation has also been improved.

* docs(module-http): Update documentation for http module

This commit updates the documentation in `README.md` for the http module. It includes an introduction to the module, a concepts section highlighting key concepts, a sequence diagram for http request execution, examples, and improved documentation for configuring http module and working with http clients. The formatting of the documentation has also been improved.

---------

Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>