feat(module-bookmark): rework bookmark module
Needed rewrite of the bookmark module to better represent the api, and provide a more robust interface for working with bookmarks. Instead of fixing the current implementation, it was decided to rework the entire module to save time and confusion in the future.

BREAKING CHANGES:
- new configurator
- new provider interface

fixup

squash

squash

squash

squash