feat(services): total rework of api interface for bookmarks. (#2354)
* Empty-Commit

* refactor: Total rework of api interface for bookmarks.

The current version misrepresents the api, and does not provide a good interface for working with bookmarks. So was decided to rework the api interface to better represent the api, and provide a more robust interface for working with bookmarks. Instead of fixing the current implementation, it was decided to rework the entire module to save time and confusion in the future.

squash

squash