feat(http): Add support for resolving URLs with a base URL that contains a path
This commit adds a new feature to the `HttpClient` class in the `@equinor/fusion-framework-module-http` package. The `HttpClient._resolveUrl` method now supports resolving URLs with a base URL that contains a path.

Before this change, when using the `HttpClient` to fetch a relative path, the base URL was not taken into account. With this change, the `HttpClient` correctly resolves the URL by appending the relative path to the base URL.