feat(module-http): Add requestValidationOperator
Add a new feature to the `@equinor/fusion-framework-module-http` package. This feature introduces the `requestValidationOperator`, a utility function that validates incoming requests against a Zod schema.

The function has options for strict validation and parsing of the request object. This operator is meant to be used as a request operator in the Fusion API framework.