Feat/module-http/validate-request-init (#2491)
* feat(module-http): Add capitalizeRequestMethodOperator

Add a new operator `capitalizeRequestMethodOperator` to ensure that the `RequestInit` method of a given request is in uppercase.

* feat(module-http): Add requestValidationOperator

Add a new feature to the `@equinor/fusion-framework-module-http` package. This feature introduces the `requestValidationOperator`, a utility function that validates incoming requests against a Zod schema.

The function has options for strict validation and parsing of the request object. This operator is meant to be used as a request operator in the Fusion API framework.

* feat(module-http): Enable capitalizeRequestMethodOperator and requestValidationOperator by default

* feat(module-http): Add remove method to ProcessOperators

* refactor(module-http): Move schema to own file

* test(http-module): Add tests for request operators

* chore(module-http): update deps

added zod

* refactor(module-http): Refactor capitalizeRequestMethodOperator to allow handle errors silently

* doc(http-module): update doc for request operators

* refactor: Capitalize HTTP request method verb to uppercase