chore(module-msal): Update @equinor/fusion-framework-module-msal to version 3.25
This commit updates the @equinor/fusion-framework-module-msal package from version 2.39 to version 3.25. The update should not require any changes to the codebase, but it is recommended to thoroughly test the application for compatibility with the new version. This update aligns with the new version of MSAL and prepares for a future major update of @equinor/fusion-framework-module-msal with breaking changes.