Rewrite bookmark (#2410)
* feat(services): total rework of api interface for bookmarks. (#2354)

* Empty-Commit

* refactor: Total rework of api interface for bookmarks.

The current version misrepresents the api, and does not provide a good interface for working with bookmarks. So was decided to rework the api interface to better represent the api, and provide a more robust interface for working with bookmarks. Instead of fixing the current implementation, it was decided to rework the entire module to save time and confusion in the future.

squash

squash

* feat(module bookmark): rewrite (#2355)

* feat(module-bookmark): rework bookmark module

Needed rewrite of the bookmark module to better represent the api, and provide a more robust interface for working with bookmarks. Instead of fixing the current implementation, it was decided to rework the entire module to save time and confusion in the future.

BREAKING CHANGES:
- new configurator
- new provider interface

* feat(react-module-bookmark): add useBookmarkProvider hook

This commit adds the `useBookmarkProvider` hook to the bookmark module. This hook allows users to retrieve the bookmark provider instance from the current modules. It enhances the bookmark management capabilities and improves the usability of the module.

feat: Add enableBookmark and additional types to bookmark module

This commit adds the `enableBookmark` function to the bookmark module, allowing users to enable bookmark functionality. It also includes additional types such as `Bookmark`, `Bookmarks`, `BookmarkData`, `BookmarkNew`, `BookmarkPayloadGenerator`, and `BookmarkProvider` to enhance the bookmark management capabilities and improve the usability of the module.

feat: Update useCurrentBookmark options

This commit modifies the `useCurrentBookmarkOptions` type in the `useCurrentBookmark.ts` file. The `payloadGenerator` property is now optional, allowing users to provide a payload generator function if needed. This change enhances the flexibility and usability of the `useCurrentBookmark` functionality.

* feat(react-app): Update bookmark module exports

This commit updates the exports in the bookmark module to use the wildcard import syntax. Instead of individually exporting specific functions and types, the module now exports everything from the '@equinor/fusion-framework-react-module-bookmark' module. This change improves the usability and flexibility of the bookmark module.

* feat(rect-components-bookmark): Make component to use provided bookmark provider

- BookmarkProvider should have provider as a parameter
- BookmarkProvider should share context of provided provider to child components
- Remove deprecated hook
- Update component to align with changes in modules

BREAKING CHANGES: BookmarkProvider now requires provider as parameter

* feat(framework-cli): Update bookmark module to use provided provider

This commit updates the bookmark module to use the provided bookmark provider. The `BookmarkProvider` component now requires the provider as a parameter and shares the context of the provided provider to its child components. The deprecated hook has been removed, and the component has been updated to align with the changes in the modules.

BREAKING CHANGES: The `BookmarkProvider` now requires the provider as a parameter.

* feat(cookbook-react-bookmark): Update cookbook to use new bookmark module

* feat(edit-bookmark): Refactor app provider query to use getAppManifest method